Split selection tests to satisfy 1000-line file length cap
Move transcript text selection tests (issue #151) from mod_tests.rs
(was 1118 lines) into selection_tests.rs (489 lines). mod_tests.rs
now 624 lines. Both under the 1000-line CI structural check limit.

Also remove dead render_transcript_text and clean up unused imports
in render_messages.rs.
